Improved inspector, added checkbox, dropdown, script titles, the ability to add and remove scripts and bug fixes
There are now 2 more UI elements, checkboxes and dropdowns.
Scripts in the inspector show their title and are grouped so the user can see which attributes are part of which scripts.
There is now a button to remove a script from the inspector.
There is now a button to add a script (it opens a dropdown from which you can select a script).
Bug fixes:
-Exiting PlayMode wouldn't reset the inspector
-Exiting PlayMode would lose the colored text on the selected object in the hierarchy