SpawnerScript, basic HoleyCastle scene, Bug fixes
SpawnerScript allows you to spawn entities at runtime. It is also the first script to utilise the dynamic inspector, since it has a vector which can be resized.
Basic Holey Castle scene currently includes the castle with walls, some trees and a basic enemy which spawn multiple times and attacks the castle.
Bugs fixed:
-Fixed the isActive checkbox not appearing in the inspector
-Fixed memory leak when loading invalid scenes
-Fixed editor closing when reading invalid scenes